Update faq.adoc
This update is almost entirely adding a space between a paragraphs end and the "Learn More:" line. Current AsciiDoc formatting says that when using the plus(+) sign to generate a new line you place the plus(+) at the end of the line *prior* to the new line. This was not the case for most of the doc so I modified it to be so. This was causing broken new lines in IE. In the process I fixed a couple broken links that were formatted wrong and removed some stray asterisks(*) that were not accomplishing what was intended by whomever placed them. Rather than fix them for intended use, I just removed them for now. Also removed a stray vertical bar(|). Applying the correct use of plus(+) as stated by the docs completely fixed all problems in IE, as for other browsers, I do not know.