Update hello_asset.adoc
I used the code from jme3test to update lines 62 and 153. I  also updated the tip on 165. This is my first time editing the wiki and I already tried editing this but did not do a pull request and I think I have to. Sorry if I messed anything up by doing it twice.