Wiki refresh (Hello Assets) (#153)
I am going through the beginner wiki pages and updating them so they are the same as the source code files in jme3test/helloworld. 

The main difference in this change is setting the DisplayStatView to false, adding normalizeLocal to the sun and then a few lines of assetManager materials. Very minor changes with the code being almost the same to start with.