Wiki refresh (Hello Physics) (#162)
I am going through the beginner wiki pages and updating them so they are the same as the source code files in jme3test/helloworld. I will do the animation page once I master it and that will be more in depth because animation spans a couple of wiki pages so I will be consolidating tweaking multiple pages. 

In this we edit the Material lines, update the RigidBodyControl lines, update the initInputs, update the private ActionListener, update the initWall method, update the initCrosshairs method, and update some comments throughout. 

jme3test/helloworld source code:
https://github.com/jMonkeyEngine/jmonkeyengine/blob/master/jme3-examples/src/main/java/jme3test/helloworld/HelloPhysics.java